pyproject.toml
[project]
name = "memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy>=3.1.3",
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
